+++ internal/storage/create.go
@@ -1,0 +1,5 @@
+package storage
+
+func (r *Repository) Create() error {
+	return nil
+}